Move classes into packages
Classes are all in the default package

Grouping classes in packages helps to standardize where similar
classes should be located

Let's create packages for some of these classes
