+++ src/main/java/pib/enums/TaskType.java
@@ -1,4 +1,4 @@
-package pib;
+package pib.enums;
 
 public enum TaskType {
     TODO, DEADLINE, EVENT
