+++ src/main/java/pib/tasks/Task.java
@@ -1,4 +1,4 @@
-package pib.Tasks;
+package pib.tasks;
 
 import pib.Ui;
